requirements.txt
requests
selenium
unstructured[all]
pillow
python-magic-bin
transformers
torch
accelerate
python-dotenv
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-google-genai
sentence-transformers
hf_xet
chainlit
rank_bm25
tiktoken
streamlit